merge: develop 브랜치와 충돌 파일 모두 해결 #108

+++ backend/app/drawing/models.py
@@ -4,7 +4,7 @@ from ..database import Base
 class DrawingTest(Base):
     __tablename__ = "drawing_tests"
 
-    drawing_id = Column(BigInteger, primary_key=True, autoincrement=True)
+    drawing_id = Column(Integer, primary_key=True, autoincrement=True)
     report_id = Column(Integer, ForeignKey("reports.report_id"))
     image_url = Column(String(255))
     risk_score = Column(Integer)
